+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 In reverse chronological order:
 
+- send velocity limits to the actuators, if they are in position mode
 - rename the node executable so that the suffixes reflect the protocl version
 - make plugins for this hardware interface to be included in [combined hardware interfaces](http://wiki.ros.org/combined_robot_hw)
 - implement joint limits (parameterised through URDF or param server)